impr: move generator logic to own class
