+++ src/main/java/com/slampvp/factory/util/Constants.java
@@ -1,7 +1,7 @@
 package com.slampvp.factory.util;
 
-public class Constants {
-    public static class Plot {
+public final class Constants {
+    public static final class Plot {
         public static final int HEIGHT = 40;
         public static final int PLOT_SIZE = 20;   // This needs to be an even number.
         public static final int ROAD_WIDTH = 3;
